Add test for metric format
